Break out grid to separate package
