+++ embed/views/playlist.js
@@ -1,6 +1,6 @@
 const html = require('choo/html')
 const imagePlaceholder = require('@resonate/svg-image-placeholder')
-const Grid = require('beta-player/build/components/grid')
+const Grid = require('@resonate/grid')
 const Playlist = require('@resonate/playlist-component')
 
 module.exports = view
