+++ packages/grid/README.md
@@ -1,0 +1,17 @@
+# @resonate/grid
+
+Render a grid element.
+
+## Install
+
+```sh
+npm i --save @resonate/grid
+```
+
+## License
+
+MIT
+
+## Author(s)
+
+- Augustin Godiscal <auggod@resonate.is>
